requirements.txt
# LangChain and MCP dependencies
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0

# MCP
mcp>=1.0.0
fastmcp>=2.0.0

# MCP server tools (for testing)
# Install with: npm install -g @modelcontextprotocol/server-filesystem

# Utilities
pyyaml>=6.0
pydantic>=2.0
python-dotenv>=1.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.21.0
